Move put account views together
